SoftwareInstaller — копия/requirements.txt
customtkinter
psutil
Pillow
pyqrcode
cairosvg
